A bug was fixed within the Game.cpp implementation that caused the program to crash whenever Keyboard events other than pressing 'W', 'A', 'S', or 'D' were pressed.
This occurred because an uninitialized variable (called direction) within the handlePlayerInput() function was being passed as a function argument if any other button was pressed.

A quick fix to this solution was to remove the variable altogether and directly call the changeDirection() function within each if statement in the handlePlayerInput() function.